Update existing samples
Make the edits for raw string literals to most content. For the programming guide content, move samples, format and update existing content.
