+++ docs/csharp/programming-guide/strings/snippets/Program.cs
@@ -1,0 +1,12 @@
+﻿// See https://aka.ms/new-console-template for more information
+
+using StringExamples;
+
+Declarations.DeclareStrings();
+Declarations.Immutability();
+
+Literals.StringLiterals();
+
+Interpolation.InterpolationExamples();
+StringCharacters.StringComponents();
+StringCharacters.BuildStrings();
