+++ samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideStrings/CS/Strings.cs
@@ -10,98 +10,6 @@ namespace CsCsrefProgrammingStrings
     {
         static void Main()
         {
-            //<Snippet1>
-            // Declare without initializing.
-            string message1;
-
-            // Initialize to null.
-            string message2 = null;
-
-            // Initialize as an empty string.
-            // Use the Empty constant instead of the literal "".
-            string message3 = System.String.Empty;
-
-            // Initialize with a regular string literal.
-            string oldPath = "c:\\Program Files\\Microsoft Visual Studio 8.0";
-
-            // Initialize with a verbatim string literal.
-            string newPath = @"c:\Program Files\Microsoft Visual Studio 9.0";
-
-            // Use System.String if you prefer.
-            System.String greeting = "Hello World!";
-
-            // In local variables (i.e. within a method body)
-            // you can use implicit typing.
-            var temp = "I'm still a strongly-typed System.String!";
-
-            // Use a const string to prevent 'message4' from
-            // being used to store another string value.
-            const string message4 = "You can't get rid of me!";
-
-            // Use the String constructor only when creating
-            // a string from a char*, char[], or sbyte*. See
-            // System.String documentation for details.
-            char[] letters = { 'A', 'B', 'C' };
-            string alphabet = new string(letters);
-            //</Snippet1>
-
-
-            //<Snippet2>
-            string s1 = "A string is more ";
-            string s2 = "than the sum of its chars.";
-
-            // Concatenate s1 and s2. This actually creates a new
-            // string object and stores it in s1, releasing the
-            // reference to the original object.
-            s1 += s2;
-
-            System.Console.WriteLine(s1);
-            // Output: A string is more than the sum of its chars.
-            //</Snippet2>
-
-
-            //<Snippet3>
-            string columns = "Column 1\tColumn 2\tColumn 3";
-            //Output: Column 1        Column 2        Column 3
-
-            string rows = "Row 1\r\nRow 2\r\nRow 3";
-            /* Output:
-              Row 1
-              Row 2
-              Row 3
-            */
-
-            string title = "\"The \u00C6olean Harp\", by Samuel Taylor Coleridge";
-            //Output: "The Æolean Harp", by Samuel Taylor Coleridge
-            //</Snippet3>
-
-            System.Console.WriteLine(columns);
-            System.Console.WriteLine(rows);
-            System.Console.WriteLine(title);
-
-
-            //PARSNIP NOTE: This snippet is far left indented to preserve formatting on multiline verbatim string
-            //<Snippet4>
-            string filePath = @"C:\Users\scoleridge\Documents\";
-            //Output: C:\Users\scoleridge\Documents\
-
-            string text = @"My pensive SARA ! thy soft cheek reclined
-                Thus on mine arm, most soothing sweet it is
-                To sit beside our Cot,...";
-            /* Output:
-            My pensive SARA ! thy soft cheek reclined
-               Thus on mine arm, most soothing sweet it is
-               To sit beside our Cot,...
-            */
-
-            string quote = @"Her name was ""Sara.""";
-            //Output: Her name was "Sara."
-            //</Snippet4>
-            System.Console.WriteLine(filePath);
-            System.Console.WriteLine(text);
-            System.Console.WriteLine(quote);
-
-
             //<Snippet5>
             string p1 = "\\\\My Documents\\My Files\\";
             string p2 = @"\\My Documents\My Files\";
@@ -117,33 +25,6 @@ namespace CsCsrefProgrammingStrings
             System.Console.WriteLine(msg);  // outputs "Eve was born in 1999"
             //</Snippet6>
 
-
-            //<Snippet7>
-            string s3 = "Visual C# Express";
-            System.Console.WriteLine(s3.Substring(7, 2));
-            // Output: "C#"
-
-            System.Console.WriteLine(s3.Replace("C#", "Basic"));
-            // Output: "Visual Basic Express"
-
-            // Index values are zero-based
-            int index = s3.IndexOf("C");
-            // index = 7
-            //</Snippet7>
-            System.Console.WriteLine("index =" + index);
-
-            //<Snippet8>
-            string s5 = "Printing backwards";
-
-            for (int i = 0; i < s5.Length; i++)
-            {
-                System.Console.Write(s5[s5.Length - i - 1]);
-            }
-            // Output: "sdrawkcab gnitnirP"
-            //</Snippet8>
-            System.Console.WriteLine();
-
-
             //<Snippet10>
             string s6 = "Battle of Hastings, 1066";
 
@@ -152,9 +33,6 @@ namespace CsCsrefProgrammingStrings
             System.Console.WriteLine(s6.ToLower());
             // outputs "battle of hastings 1066"
             //</Snippet10>
-
-
-
 
             System.Console.ReadKey();
         }
@@ -196,28 +74,6 @@ namespace CsCsrefProgrammingStrings
             Console.ReadKey();
         }
     }
-
-
-
-
-
-
-    class UsingStringBuilder
-    {
-        static void Main()
-        {
-            //-------------------------------------------------------------------------
-            //<Snippet15>
-            System.Text.StringBuilder sb = new System.Text.StringBuilder("Rat: the ideal pet");
-            sb[0] = 'C';
-            System.Console.WriteLine(sb.ToString());
-            System.Console.ReadLine();
-
-            //Outputs Cat: the ideal pet
-            //</Snippet15>
-        }
-    }
-
 
     //-------------------------------------------------------------------------
     //<Snippet16>
@@ -371,51 +227,6 @@ namespace CsCsrefProgrammingStrings
         }
     }
 
-
-    //-------------------------------------------------------------------------
-    class StringBuilderExample2
-    {
-        static void Main()
-        {
-            //<Snippet20>
-            static void Main()
-            {
-                string str = "hello";
-                string nullStr = null;
-                string emptyStr = String.Empty;
-
-                string tempStr = str + nullStr;
-                // Output of the following line: hello
-                Console.WriteLine(tempStr);
-
-                bool b = (emptyStr == nullStr);
-                // Output of the following line: False
-                Console.WriteLine(b);
-
-                // The following line creates a new empty string.
-                string newStr = emptyStr + nullStr;
-
-                // Null strings and empty strings behave differently. The following
-                // two lines display 0.
-                Console.WriteLine(emptyStr.Length);
-                Console.WriteLine(newStr.Length);
-                // The following line raises a NullReferenceException.
-                //Console.WriteLine(nullStr.Length);
-
-                // The null character can be displayed and counted, like other chars.
-                string s1 = "\x0" + "abc";
-                string s2 = "abc" + "\x0";
-                // Output of the following line: * abc*
-                Console.WriteLine("*" + s1 + "*");
-                // Output of the following line: *abc *
-                Console.WriteLine("*" + s2 + "*");
-                // Output of the following line: 4
-                Console.WriteLine(s2.Length);
-            }
-            //</Snippet20>
-        }
-    }
-
     //-------------------------------------------------------------------------
     //<Snippet21>
     class StringSearch
@@ -555,26 +366,7 @@ namespace CsCsrefProgrammingStrings
       The quick brown cat jumped over the fence
     */
     //</Snippet24>
-    //-------------------------------------------------------------------------
 
-    class StringReferenceDemo
-    {
-        static void Main()
-        {
-            // <snippet25>
-            string s1 = "Hello ";
-            string s2 = s1;
-            s1 += "World";
-
-            System.Console.WriteLine(s2);
-            //Output: Hello
-            //</snippet25>
-
-            System.Console.ReadKey();
-        }
-    }
-
-    //-------------------------------------------------------------------------
     //<snippet26>
     class FormatString
     {
@@ -602,31 +394,6 @@ namespace CsCsrefProgrammingStrings
         }
     }
     //</snippet26>
-    //-------------------------------------------------------------------------
-
-    class StringNullReference
-    {
-    static void Main()
-    {
-        //<snippet27>
-        string question = "hOW DOES mICROSOFT wORD DEAL WITH THE cAPS lOCK KEY?";
-        System.Text.StringBuilder sb = new System.Text.StringBuilder(question);
-
-        for (int j = 0; j < sb.Length; j++)
-        {
-            if (System.Char.IsLower(sb[j]) == true)
-                sb[j] = System.Char.ToUpper(sb[j]);
-            else if (System.Char.IsUpper(sb[j]) == true)
-                sb[j] = System.Char.ToLower(sb[j]);
-        }
-        // Store the new string.
-        string corrected = sb.ToString();
-        System.Console.WriteLine(corrected);
-        // Output: How does Microsoft Word deal with the Caps Lock key?
-        //</snippet27>
-    }
-    }
-    //-------------------------------------------------------------------------
 
     //<snippet28>
     class ReplaceSubstrings
